ACME mock server refactoring
