+++ tests/integration/test_acme_tls/mocks/cert.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDnzCCAoegAwIBAgIUcD5Cx5SX7BXjxEKsygJQ8m8vlWowDQYJKoZIhvcNAQEL
+BQAwdDELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVBAcM
+DVNhbiBGcmFuY2lzY28xEzARBgNVBAoMCkNsaWNrSG91c2UxIzAhBgNVBAMMGmtv
+bnN0YW50aW4tY2xpY2tob3VzZTo4NDQzMB4XDTI1MTExMDE3NDQ1NVoXDTI2MTEx
+MDE3NDQ1NVowdDELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAU
+BgNVBAcMDVNhbiBGcmFuY2lzY28xEzARBgNVBAoMCkNsaWNrSG91c2UxIzAhBgNV
+BAMMGmtvbnN0YW50aW4tY2xpY2tob3VzZTo4NDQzMIIBIjANBgkqhkiG9w0BAQEF
+AAOCAQ8AMIIBCgKCAQEAyVJ9rceI68njrNyhvOc4gJnR1sTXHXAgWqCq0zTQVX4j
+DrwfFsL6pCPS34jesqAIXU8G/VXh0RJZkOgNhfGtQIhrcMVI5BkA/iLoO+YMGKQe
+cwOMRdV0ANOemERjuf/5VRFw1Y3XMjGlQMDNtq0MfjTaIxxPNt5XZkwAH2siHGpw
+D4Krzlec0OZlqATUWKglkZiWheM39VGzKp2yt6HSxgTS3xF0PEORuOVBCidF+2oZ
+inMOrPl3gB8mRrzrE3WdfXx2Iov5eMT9gYq++KyiIMU6OLY7PW+JO5wPC0ETurVw
+ECdazbmENRzt5Zcgraz5NhpfDjhLsjLFkgVExQFEWQIDAQABoykwJzAlBgNVHREE
+HjAcghprb25zdGFudGluLWNsaWNraG91c2U6ODQ0MzANBgkqhkiG9w0BAQsFAAOC
+AQEAgkRx0Gxn/QBjiTlo+uYZO8E9Ehj5mFuxWofRqaW02r8QDOV+Qe3SK7/tbV4v
+kqjNoY00UnpzCRyi8dpF+36bm5dv3odIAXpluj2ZsPJ5JsJ1xsUhXD7odcbW+Iyd
+EhRrAyrQXLNKieLV8y6kwvuXzi9FW7kjXqlCymh+6W0AAi2xFkl+nOMzJ2P/yccX
+lY7CGDU95qbSM036uOjpGSCMULtAheQc0rJsEyS8IMiaFACH711AEfoWjzvtH0ZS
++uyG/xGedh/cpkYRGrEeAOHJaqH+27f0o0xWUaCBf414hMl6oYy6OrDo/lL3W6wF
+UF+cESBWrmn75owXsp8WWFlWXA==
+-----END CERTIFICATE-----
