+++ tests/integration/test_acme_tls/mocks/certificates.py
@@ -1,0 +1,81 @@
+import datetime
+
+from cryptography import x509
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import hashes, serialization
+from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.x509.oid import NameOID
+
+def generate_self_signed_cert(hostname):
+    key = rsa.generate_private_key(
+        public_exponent=65537, key_size=2048, backend=default_backend()
+    )
+    utc_now = datetime.datetime.now(datetime.timezone.utc)
+
+    # Generate self-signed certificate
+    subject = issuer = x509.Name(
+        [
+            x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
+            x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "California"),
+            x509.NameAttribute(NameOID.LOCALITY_NAME, "San Francisco"),
+            x509.NameAttribute(NameOID.ORGANIZATION_NAME, "ClickHouse"),
+            x509.NameAttribute(NameOID.COMMON_NAME, hostname),
+        ]
+    )
+    cert = (
+        x509.CertificateBuilder()
+        .subject_name(subject)
+        .issuer_name(issuer)
+        .public_key(key.public_key())
+        .serial_number(x509.random_serial_number())
+        .not_valid_before(utc_now)
+        .not_valid_after(utc_now + datetime.timedelta(days=365))
+        .add_extension(
+            x509.SubjectAlternativeName([x509.DNSName(hostname)]),
+            critical=False,
+        )
+        .sign(key, hashes.SHA256(), default_backend())
+    )
+
+    # Serialize key and certificate to memory
+    private_key_pem = key.private_bytes(
+        encoding=serialization.Encoding.PEM,
+        format=serialization.PrivateFormat.TraditionalOpenSSL,
+        encryption_algorithm=serialization.NoEncryption(),
+    )
+    cert_pem = cert.public_bytes(serialization.Encoding.PEM)
+
+    return private_key_pem, cert_pem
+
+
+def generate_certificate(csr, issuer_key, days_valid=365):
+    private_key = serialization.load_pem_private_key(issuer_key, password=None)
+
+    subject = csr.subject
+    issuer = x509.Name(
+        [
+            x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
+            x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "California"),
+            x509.NameAttribute(NameOID.LOCALITY_NAME, "San Francisco"),
+            x509.NameAttribute(NameOID.ORGANIZATION_NAME, "ClickHouse"),
+            x509.NameAttribute(NameOID.COMMON_NAME, "Integration tests"),
+        ]
+    )
+
+    utc_now = datetime.datetime.now(datetime.timezone.utc)
+
+    # Build the certificate
+    cert_builder = (
+        x509.CertificateBuilder()
+        .subject_name(subject)
+        .issuer_name(issuer)
+        .public_key(csr.public_key())
+        .serial_number(x509.random_serial_number())
+        .not_valid_before(utc_now)
+        .not_valid_after(utc_now + datetime.timedelta(days=days_valid))
+        .add_extension(x509.BasicConstraints(ca=False, path_length=None), critical=True)
+    )
+
+    # Sign the certificate
+    cert = cert_builder.sign(private_key=private_key, algorithm=hashes.SHA256())
+    return cert
